v0.5.1.0-alpha stuff.

+++ CPPZombs/Project1/Particles.h
@@ -49,7 +49,7 @@ public:
 		if (color.a == 255 && tTime == startTime)
 			return game->DrawCircle(pos, color, radius);
 		glDepthMask(GL_FALSE);
-		game->DrawCircle(pos, RGBA(color.r, color.g, color.b, color.a * (1.f - (tTime - startTime) / duration)), radius);
+		game->DrawCircle(pos, RGBA(color.r, color.g, color.b, static_cast<byte>(color.a * (1.f - (tTime - startTime) / duration))), radius);
 		glDepthMask(GL_TRUE);
 	}
 };
